Changed textures to 32-bit-float-per-color in order to have appropriate precision necessary for flow calculations. Could maybe be reduced in future if required for speed, as float-32 is most precise but most expensive.

+++ Flow/Assets/DirtShader.shader
@@ -49,10 +49,10 @@
 				return o;
 			}
 			
-			fixed4 frag (v2f i) : SV_Target
+			float4 frag (v2f i) : SV_Target
 			{
 				// sample the texture
-				fixed4 pixel = tex2D(_MainTex, i.uv);
+				float4 pixel = tex2D(_MainTex, i.uv);
 
 				return pixel;
 			}
