+++ Flow/Assets/HeightShader.shader
@@ -50,14 +50,14 @@
 				return o;
 			}
 			
-			fixed4 frag (v2f i) : SV_Target
+			float4 frag (v2f i) : SV_Target
 			{
 				// Sum the heights of all elements
-				fixed4 water = tex2D(_WaterTex, i.uv);
-				fixed4 mud = tex2D(_MudTex, i.uv);
-				fixed4 dirt = tex2D(_DirtTex, i.uv);
+				float4 water = tex2D(_WaterTex, i.uv);
+				float4 mud = tex2D(_MudTex, i.uv);
+				float4 dirt = tex2D(_DirtTex, i.uv);
 
-				return fixed4((dirt.r + mud.r + water.r) / _NumElements, 0, 0, 1);
+				return float4((dirt.r + mud.r + water.r) / _NumElements, 0, 0, 1);
 			}
 			ENDCG
 		}
